Move JetsSeatMapService tests for handler functions to service.handlers.test.js, tests for getSeatMapData to service.getSeatMapData.test.js (will be committed as part of a separate PR).
